Be paranoid in D-Bus signal callbacks
Added logic in the D-Bus signal callbacks to be paranoid about the
input to the callback, as it became clear from debugging that having
two different libraries use g_dbus_connection_signal_subscribe in
the same process results in the separately registered callbacks all
getting called with the superset of the filtered signal requests.
At best this leads to confusion with respect to logging/debugging,
at worst it opens the door to tripping over unexpected inputs in the
callbacks.

Bug-AGL: SPEC-4182

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I86a804917c5f31f4c94e3e642728f7e0df5dd2e4